ENCODE THE MESSAGE
You are a secret agent on a mission, and need to decode various messages using the ASCII values of characters. Due to
resource constraints, you can send only those messages with a total ASCII sum  equal to or  less than a specified threshold
N. You need to remove characters from the end of the string to maintain the sum accordingly. Your task is to find and return the
decoded string, or return an empty string, if it is not possible to meet the conditions

Input Specification:
input1: A string value S, representing the encoded message.
input2: An integer value N, representing the threshold value.
Output Specification:
Return a string value representing the decoded string, or an empty string if it is not possible to meet the condition.