Create NUMBER OF COMBINATIONS LEADING TO A PRODUCT
Problem Statement:
You are given an array arr and a product m. Your task is to find the number of possible unique triplets whose product of
elements is m. 
Input Format:
The first line contains the integer, n
The second line contains space seperated integers of the array, arr
The third line contains the product m.
The input will be read from the STDIN by the candidate
Output Format:
The output consists of a single integer, i.e. the count of unique triplets having product m.